+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+### Added
+
+- Viewer: long press on rating quick action for quicker rating
+
 ### Changed
 
 - Viewer: allow setting default outside video player